Refactor MCH program management moving it to care pannel: Remove unused program forms and related components from mch
