+++ packages/esm-billing-app/translations/am.json
@@ -239,6 +239,7 @@
   "Phone Number": "ስልክ ቁጥር",
   "pleaseSelectItems": "ጥያቄ ለማቅረብ እባክዎ መስመር ዕቃዎችን ይምረጡ",
   "policyNumber": "የፖሊሲ ቁጥር",
+  "preAuth": "Pre-Authorization",
   "preauthRequest": "ቅድመ ፍቃድ ጥያቄዎች",
   "preAuthRequests": "ቅድመ ፍቃድ ጥያቄዎች",
   "preAuthRequets": "ቅድመ ፍቃድ ጥያቄዎች",
