+++ packages/esm-billing-app/translations/en.json
@@ -239,6 +239,7 @@
   "Phone Number": "Phone Number",
   "pleaseSelectItems": "Please select line items to raise a claim",
   "policyNumber": "Policy number",
+  "preAuth": "Pre-Authorization",
   "preauthRequest": "Preauth requests",
   "preAuthRequests": "Pre-auth requests",
   "preAuthRequets": "Pre-Auth Requests",
