+++ packages/esm-care-panel-app/src/care-programs/care-programs.component.tsx
@@ -1,26 +1,35 @@
-import React, { useCallback, useMemo } from 'react';
 import {
-  InlineLoading,
   Button,
   DataTable,
+  DataTableSkeleton,
+  InlineLoading,
+  OverflowMenu,
+  OverflowMenuItem,
   Table,
-  TableHeader,
-  TableRow,
-  TableHead,
   TableBody,
   TableCell,
   TableContainer,
+  TableHead,
+  TableHeader,
+  TableRow,
+  Tag,
   Tile,
-  DataTableSkeleton,
 } from '@carbon/react';
 import { Close, DocumentAdd } from '@carbon/react/icons';
-import { CardHeader, EmptyState, launchStartVisitPrompt, ErrorState } from '@openmrs/esm-patient-common-lib';
-import { useTranslation } from 'react-i18next';
-import { PatientCarePrograms, useCarePrograms } from '../hooks/useCarePrograms';
 import { formatDate, launchWorkspace, restBaseUrl, useLayoutType, useVisit } from '@openmrs/esm-framework';
+import { CardHeader, EmptyState, ErrorState, launchStartVisitPrompt } from '@openmrs/esm-patient-common-lib';
 import capitalize from 'lodash/capitalize';
+import React, { useCallback, useMemo } from 'react';
+import { useTranslation } from 'react-i18next';
 import { mutate } from 'swr';
+import { PatientCarePrograms, useCarePrograms } from '../hooks/useCarePrograms';
 
+import {
+  getProgramForms,
+  launchDeleteProgramDialog,
+  launchProgramForm,
+  usePatientEnrolledPrograms,
+} from './care-program.resource';
 import styles from './care-programs.scss';
 
 type CareProgramsProps = {
@@ -30,7 +39,13 @@ type CareProgramsProps = {
 const CarePrograms: React.FC<CareProgramsProps> = ({ patientUuid }) => {
   const { t } = useTranslation();
   const { currentVisit } = useVisit(patientUuid);
-  const { carePrograms, isLoading, isValidating, error, mutateEligiblePrograms } = useCarePrograms(patientUuid);
+  const { eligibleCarePrograms, isLoading, isValidating, error, mutateEligiblePrograms } = useCarePrograms(patientUuid);
+  const {
+    enrollments,
+    isLoading: isLoadingEnrollments,
+    error: enrollmentsError,
+    mutate: mutateEnrollments,
+  } = usePatientEnrolledPrograms(patientUuid);
   const isTablet = useLayoutType() === 'tablet';
 
   const handleMutations = useCallback(() => {
@@ -75,8 +90,55 @@ const CarePrograms: React.FC<CareProgramsProps> = ({ patientUuid }) => {
   );
 
   const rows = useMemo(
-    () =>
-      carePrograms.map((careProgram) => {
+    () => [
+      ...enrollments.map((enrollment) => {
+        const forms = getProgramForms(enrollment.program.uuid);
+
+        return {
+          id: enrollment.program.uuid,
+          programName: enrollment.program.name,
+          status: (
+            <div className={styles.careProgramButtonContainer}>
+              <Tag type="green">Enrolled</Tag>
+              <OverflowMenu aria-label="overflow-menu" flipped>
+                {forms.map((form) => (
+                  <OverflowMenuItem
+                    key={form.formUuId}
+                    itemText={form.formName}
+                    onClick={() => {
+                      currentVisit
+                        ? launchWorkspace('patient-form-entry-workspace', {
+                            workspaceTitle: form.formName,
+                            mutateForm: () => {
+                              mutateEnrollments();
+                              mutateEligiblePrograms();
+                            },
+                            formInfo: {
+                              encounterUuid: '',
+                              formUuid: form.formUuId,
+                              // additionalProps: { enrollmenrDetails: careProgram.enrollmentDetails ?? {} },
+                            },
+                          })
+                        : launchStartVisitPrompt();
+                    }}
+                  />
+                ))}
+                <OverflowMenuItem
+                  itemText={t('edit', 'Edit')}
+                  onClick={() =>
+                    launchProgramForm(enrollment.program.uuid, patientUuid, enrollment, () => mutateEnrollments())
+                  }
+                />
+                <OverflowMenuItem
+                  itemText={t('delete', 'Delete')}
+                  onClick={() => launchDeleteProgramDialog(enrollment.uuid, patientUuid)}
+                />
+              </OverflowMenu>
+            </div>
+          ),
+        };
+      }),
+      ...eligibleCarePrograms.map((careProgram) => {
         return {
           id: `${careProgram.uuid}`,
           programName: careProgram.display,
@@ -96,7 +158,13 @@ const CarePrograms: React.FC<CareProgramsProps> = ({ patientUuid }) => {
                 className="cds--btn--sm cds--layout--size-sm"
                 kind={careProgram.enrollmentStatus == 'active' ? 'danger--ghost' : 'ghost'}
                 iconDescription="Dismiss"
-                onClick={() => handleCareProgramClick(careProgram)}
+                // onClick={() => handleCareProgramClick(careProgram)}
+                onClick={() =>
+                  launchProgramForm(careProgram.uuid, patientUuid, undefined, () => {
+                    mutateEnrollments();
+                    mutateEligiblePrograms();
+                  })
+                }
                 renderIcon={careProgram.enrollmentStatus == 'active' ? Close : DocumentAdd}>
                 {careProgram.enrollmentStatus == 'active' ? 'Discontinue' : 'Enroll'}
               </Button>
@@ -104,7 +172,8 @@ const CarePrograms: React.FC<CareProgramsProps> = ({ patientUuid }) => {
           ),
         };
       }),
-    [carePrograms, handleCareProgramClick],
+    ],
+    [enrollments, eligibleCarePrograms, t, currentVisit, mutateEnrollments, patientUuid, mutateEligiblePrograms],
   );
 
   const headers = [
@@ -126,7 +195,7 @@ const CarePrograms: React.FC<CareProgramsProps> = ({ patientUuid }) => {
     return <ErrorState headerTitle={t('errorCarePrograms', 'Care programs')} error={error} />;
   }
 
-  if (carePrograms.length === 0) {
+  if (eligibleCarePrograms.length === 0) {
     return <EmptyState headerTitle={t('careProgram', 'Care program')} displayText={t('careProgram', 'care program')} />;
   }
