+++ packages/esm-express-workflow-app/src/components/mch/index.tsx
@@ -1,10 +1,8 @@
 import { getSyncLifecycle } from '@openmrs/esm-framework';
 
 import { moduleName } from '../../constants';
-import MCHDashboard from './dashboard.component';
-import Programs from './programs/programs.component';
-import ProgramForm from './programs/program.workspace';
 import { createLeftPanelLink } from '../../shared/dashboard-link/dashboard-link.component';
+import MCHDashboard from './dashboard.component';
 
 const options = {
   featureName: 'express-workflow',
@@ -19,6 +17,3 @@ export const mchLeftPanelLink = getSyncLifecycle(
   }),
   options,
 );
-
-export const mchProgramManagement = getSyncLifecycle(Programs, options);
-export const mchProgramForm = getSyncLifecycle(ProgramForm, options);
