--- packages/esm-express-workflow-app/src/components/mch/programs/program.scss
@@ -1,45 +1,0 @@
-@use '@carbon/type';
-@use '@carbon/layout';
-@use '@carbon/colors';
-
-.heading {
-  @include type.type-style('heading-compact-01');
-  margin: layout.$spacing-05 0 layout.$spacing-05;
-}
-
-.form {
-  display: flex;
-  flex-direction: column;
-  justify-content: space-between;
-  width: 100%;
-  height: 100%;
-}
-
-.grid {
-  margin: layout.$spacing-05 layout.$spacing-05;
-  padding: layout.$spacing-05 0 0 0;
-}
-
-.button {
-  display: flex;
-  align-content: flex-start;
-  align-items: baseline;
-  min-width: 50%;
-}
-
-.buttonSet {
-  display: flex;
-  justify-content: space-between;
-  width: 100%;
-}
-
-.sectionHeader {
-  @include type.type-style('heading-02');
-}
-
-.datePickerInput span,
-.datePickerInput div,
-.datePickerInput input,
-.datePickerInput {
-  min-width: 100%;
-}
