+++ packages/esm-express-workflow-app/src/routes.json
@@ -269,12 +269,6 @@
       "order": 0
     },
     {
-      "name": "mch-program-management",
-      "component": "mchProgramManagement",
-      "slot": "mch-program-management-slot",
-      "order": 0
-    },
-    {
       "name": "ewf-encounter-details-dashboard",
       "component": "encounterDetails",
       "slot": "ewf-clinical-encounter-slot",
@@ -291,12 +285,6 @@
       "type": "workspace",
       "canMaximize": true,
       "canHide": true
-    },
-    {
-      "name": "mch-program-form-workspace",
-      "component": "mchProgramForm",
-      "title": "Program Form",
-      "type": "workspace"
     }
   ],
   "modals": [
