+++ packages/esm-express-workflow-app/translations/am.json
@@ -6,7 +6,6 @@
   "attachments": "አባሪዎች",
   "awaitingConsultation": "ምክር እየጠበቀ ያለ",
   "awaitingInvestigation": "ምርመራ እየጠበቀ ያለ",
-  "awaitingTriage": "ምደባ (ትሪያጅ) እየጠበቀ ያለ",
   "back": "ተመለስ",
   "birthDate": "የትውልድ ቀን",
   "call": "ይደውሉ",
@@ -20,10 +19,8 @@
   "checkingEligibility": "ብቁነት እየተረጋገጠ ነው",
   "checkInPatientWorkspaceTitle": "ሕመምተኛ ይመዝገቡ",
   "chooseIdentifierType": "የመለያ ዓይነት ይምረጡ",
-  "cilentsPatientInWaiting": "ተገልጋዮች/ሕሙማን በጥበቃ ላይ",
   "clearAll": "ሁሉንም አጽዳ",
   "clientRegistry": "የደንበኞች መዝገብ ማረጋገጫ",
-  "clientsPatientAttended": "ተስተናገዱ/የታዩ ተገልጋዮች/ሕሙማን",
   "clinicalEncounter": "ክሊኒካዊ ግንኙነት/ምርመራ",
   "comingFrom": "የመጡበት ቦታ",
   "confirmationTxt": "ከኦቲፒ በፊት የስልክ ቁጥሩን ያረጋግጡ",
@@ -57,7 +54,6 @@
   "invalidPhoneNumber": "እባክዎ ትክክለኛ የስልክ ቁጥር ያስገቡ",
   "investigationComplete": "ምርመራ ተጠናቋል",
   "laboratoryOrders": "የላቦራቶሪ ትዕዛዞች",
-  "launchProgramForm": "{{program}} ቅጽ ይክፈቱ",
   "loadingPatient": "ሕመምተኛ እየተጫነ ነው...",
   "loadingQueueEntries": "የጥበቃ ዝርዝር እየተጫነ ነው...",
   "loadingQueues": "የጥበቃ መስመሮች እየተጫኑ ነው...",
@@ -82,9 +78,11 @@
   "otpVerification": "ኦቲፒ ማረጋገጫ",
   "otpVerificationError": "ኦቲፒ በማረጋገጥ ላይ ስህተት",
   "pastMedications": "ያለፉ መድኃኒቶች",
+  "patientAttended": "Patient attended",
   "patientExamination": "የሕመምተኛ ምርመራ",
   "patientHistory": "የሕመምተኛ ታሪክ",
   "patientManagement": "የሕመምተኛ አስተዳደር",
+  "patientsAttendedTo": "Patients attended to",
   "patientSummary": "የሕመምተኛ ማጠቃለያ",
   "phoneNumber": "ስልክ ቁጥር",
   "previousPage": "የቀደመ ገጽ",
