+++ packages/esm-express-workflow-app/translations/en.json
@@ -6,7 +6,6 @@
   "attachments": "Attachments",
   "awaitingConsultation": "Awaiting consultation",
   "awaitingInvestigation": "Awaiting Investigation",
-  "awaitingTriage": "Awaiting Triage",
   "back": "Back",
   "birthDate": "Birth Date",
   "call": "Call",
@@ -20,10 +19,8 @@
   "checkingEligibility": "",
   "checkInPatientWorkspaceTitle": "Check in patient",
   "chooseIdentifierType": "Choose identifier type",
-  "cilentsPatientInWaiting": "Cilents/Patient in waiting",
   "clearAll": "Clear All",
   "clientRegistry": "Client registry verification",
-  "clientsPatientAttended": "Clients/Patient attended to",
   "clinicalEncounter": "Clinical Encounter",
   "comingFrom": "Coming from",
   "confirmationTxt": "Verify the phone number before OTP",
@@ -57,7 +54,6 @@
   "invalidPhoneNumber": "Please enter a valid phone number",
   "investigationComplete": "Investigation complete",
   "laboratoryOrders": "Laboratory Orders",
-  "launchProgramForm": "Launch {{program}} Form",
   "loadingPatient": "Loading patient...",
   "loadingQueueEntries": "Loading queue entries...",
   "loadingQueues": "Loading queues...",
@@ -82,9 +78,11 @@
   "otpVerification": "OTP Verification",
   "otpVerificationError": "Error Verifying OTP",
   "pastMedications": "Past Medications",
+  "patientAttended": "Patient attended",
   "patientExamination": "Patient Examination",
   "patientHistory": "Patient History",
   "patientManagement": "Patient Management",
+  "patientsAttendedTo": "Patients attended to",
   "patientSummary": "Patient Summary",
   "phoneNumber": "Phone number",
   "previousPage": "Previous page",
