+++ packages/esm-patient-clinical-view-app/translations/am.json
@@ -5,6 +5,7 @@
   "activeCases": "ገባሪ ጉዳዮች",
   "add": "ጨምር",
   "addCase": "ጉዳይ ጨምር",
+  "addFirstDataPoint": "Add first data point",
   "addPeer": "እኩያ ጨምር",
   "addPNSContact": "የPNS እውቂያ ጨምር",
   "addRelationship": "ግንኙነት ጨምር",
@@ -18,6 +19,7 @@
   "all": "ሁሉም",
   "ancVisitNumber": "የANC ጉብኝት ቁጥር",
   "antenatalCare": "ቅድመ ወሊድ እንክብካቤ",
+  "average": "Average",
   "babyGivenVitaminD": "ለህፃኑ ቫይታሚን ኬ ተሰጥቷል",
   "baselineHivStatus": "የመነሻ የኤችአይቪ ሁኔታ",
   "baselineInformation": "የመነሻ መረጃ",
@@ -27,9 +29,9 @@
   "bloodGroup": "የደም አይነት",
   "bloodLoss": "የደም ስርጭት ተደርጓል",
   "bloodTransfusion": "የደም ስርጭት",
+  "cancel": "Cancel",
   "caseEncounter": "የጉዳይ አስተዳደር ግኝቶች",
   "causeOfDeath": "የሞት መንስኤ",
-  "cervicalDilation": "የማህፀን በር መስፋፋት በሴ.ሜ",
   "chronicDisease": "ሥር የሰደደ በሽታ",
   "clientEligibility": "ብቁ",
   "clinicalEncounter": "ክሊኒካዊ ግኝት",
@@ -41,7 +43,6 @@
   "contactList": "የእውቂያ ዝርዝር",
   "contactTracingForm": "የእውቂያ ፍለጋ ቅጽ",
   "contactType": "የእውቂያ አይነት",
-  "contractionLevel": "Contraction level",
   "date": "ቀን",
   "dateDate": "ቀን እና ሰዓት",
   "dateOfDelivery": "የወሊድ ቀን",
@@ -52,7 +53,6 @@
   "deliveryDate": "የወሊድ ቀን",
   "deliveryOutcome": "የወሊድ ውጤት",
   "department": "ክፍል",
-  "descentOfHead": "የጭንቅላት መውረድ",
   "disabilities": "አካል ጉዳተኝነት",
   "discard": "አስወግድ",
   "dischargeDate": "የመልቀቂያ ቀን",
@@ -81,18 +81,18 @@
   "failedDeleting": "ሊሰረዝ አልቻለም",
   "familyContacts": "የቤተሰብ እውቂያዎች",
   "feedingOrders": "የአመጋገብ ትዕዛዞች",
-  "fetalHeartRate": "የፅንስ የልብ ምት",
+  "fieldRequired": "This field is required",
   "filterByForm": "በቅጽ አጣራ",
   "filterByStatue": "በሁኔታ አጣራ",
   "filterByStatus": "በሁኔታ አጣራ",
   "finalDiagnosis": "የመጨረሻ ምርመራ",
   "finalDischargeDiagnosis": "የመጨረሻ የመልቀቂያ ምርመራ",
   "finalOutcome": "የመጨረሻ ውጤት",
-  "foetalHeartRate": "Foetal Heart Rate",
   "foetalPresentation": "የፅንስ አቀራረብ",
   "followUpDate": "ቀጣይ የክትትል ቀን",
   "formName": "የቅጽ ስም",
   "gestationalSize": "የእርግዝና መጠን",
+  "graphView": "Graph View",
   "hivPatientSummary": "የኤችአይቪ ታካሚ ማጠቃለያ",
   "hivStatus": "የኤችአይቪ ሁኔታ",
   "hivTestResults": "የኤችአይቪ ሁኔታ",
@@ -111,14 +111,15 @@
   "ipvOutCome": "IPV ውጤት",
   "ipvQuestions": "IPV ጥያቄዎች",
   "labourAndDelivery": "ምጥ እና ወሊድ",
+  "latest": "Latest",
   "listingDate": "የመዘገብ ቀን",
   "livingWithClient": "ከታካሚው ጋር የሚኖሩ",
   "loading": "በመጫን ላይ...",
+  "loadingData": "Loading data...",
   "managePeers": "እኩያዎችን አስተዳድር",
   "maternalCondition": "የእናት ሁኔታ",
   "maternalSummary": "የእናት ማጠቃለያ",
   "medicalHistory": "የህክምና ታሪክ",
-  "membraneAmnioticFluidAndMoulding": "Membrane Amniotic Fluid & Moulding",
   "missedAppointmentDate": "ቀጠሮ ያመለጠበት ቀን",
   "modeOfDelivery": "የወሊድ ዘዴ",
   "motherGeneralCondition": "አጠቃላይ ሁኔታ",
@@ -129,14 +130,16 @@
   "nextVisitDate": "የሚቀጥለው የጉብኝት ቀን",
   "no": "አይ",
   "noContactToDisplay": "ለዚህ ታካሚ የሚታይ የእውቂያ ዳታ የለም።",
+  "noDataAvailable": "No data available for this graph",
   "noEncounters": "ምንም ግኝቶች አልተገኙም",
   "noEncounterToDisplay": "ለዚህ ታካሚ የሚታዩ ግኝቶች የሉም።",
   "noObservationsFound": "ምንም ምልከታዎች አልተገኙም",
+  "noPatientSelected": "No patient selected",
   "operatingDoctor": "የሚያሰራ ሀኪም",
   "otherSubstanceAbuse": "ሌላ የሱስ አላግባብ መጠቀም",
   "partnerStatus": "የባልደረባ የኤችአይቪ ሁኔታ",
   "partograph": "ፓርቶግራፍ",
-  "partographData": "ዋና ክፍሎች",
+  "partography": "Partography",
   "peerCalendar": "የእኩያ ካሌንደር",
   "pending": "በመጠባበቅ ላይ",
   "pendingDocumentations": "በመጠባበቅ ላይ ያለ ሰነድ",
@@ -166,8 +169,10 @@
   "save": "አስቀምጥ",
   "savedRelationship": "ግንኙነት በተሳካ ሁኔታ ተቋርጧል",
   "selectClinic": "ክሊኒክ ምረጥ",
+  "selectPatientMessage": "Please select a patient to view partography data.",
   "sex": "ፆታ",
   "sexualAssault": "3.እሱ/እሷ sexually ምቾት ያልተሰማሽን/ያልተሰማህን ነገር እንድታደርጊ/እንድታደርግ አስገድዶሽ/አስገድዶሃል?",
+  "showingResults": "Showing {{start}}-{{end}} of {{total}} {{itemType}}",
   "smoking": "ማጨስ",
   "smokingDuration": "የማጨስ ጊዜ",
   "socialHistory": "ማህበራዊ ታሪክ",
@@ -177,7 +182,7 @@
   "successfullyDeleted": "በተሳካ ሁኔታ ተሰርዟል",
   "surgicalHistory": "የቀዶ ጥገና ታሪክ",
   "surgicalSummary": "የቀዶ ጥገና ማጠቃለያ",
-  "tabList": "Tab List",
+  "tableView": "Table View",
   "tbScreening": "የቲቢ ምርመራ ውጤት",
   "testApproach": "አቀራረብ",
   "testDate": "የምርመራ ቀን",
@@ -187,7 +192,7 @@
   "testStrategy": "ስትራቴጂ",
   "therapiesPrescribed": "የታዘዙ ህክምናዎች",
   "threatened": "2. እሱ/እሷ አንቺን / አንተን ለመጉዳት አስፈራርቶሽ/አስፈራርቶሃል?",
-  "timeRecorded": "የተመዘገበበት ሰዓት",
+  "time": "Time",
   "total": "ጠቅላላ",
   "totalPeers": "ጠቅላላ እኩያዎች",
   "traceContact": "እውቂያ ፈልግ",
@@ -200,5 +205,6 @@
   "viewBeds": "አልጋዎችን እይ",
   "viewEncounter": "እይ",
   "visitDate": "የምርመራ ቀን",
+  "yearsOld": "years old",
   "yes": "አዎ"
 }
