+++ packages/esm-patient-clinical-view-app/translations/sw.json
@@ -5,6 +5,7 @@
   "activeCases": "Active cases",
   "add": "Add",
   "addCase": "Add case",
+  "addFirstDataPoint": "Add first data point",
   "addPeer": "Add Peer",
   "addPNSContact": "Add PNS Contact",
   "addRelationship": "Add relationship",
@@ -18,6 +19,7 @@
   "all": "All",
   "ancVisitNumber": "ANC Visit Number",
   "antenatalCare": "Antenatal Care",
+  "average": "Average",
   "babyGivenVitaminD": "Baby Given Vitamin K",
   "baselineHivStatus": "Baseline HIV Status",
   "baselineInformation": "Baseline Information",
@@ -27,9 +29,9 @@
   "bloodGroup": "Blood Group",
   "bloodLoss": "Blood Transfusion Done",
   "bloodTransfusion": "Blood Transfusion",
+  "cancel": "Cancel",
   "caseEncounter": "Case management encounters",
   "causeOfDeath": "Cause of Death",
-  "cervicalDilation": "Cervical Dilation cm",
   "chronicDisease": "Chronic Disease",
   "clientEligibility": "Eligible",
   "clinicalEncounter": "Clinical Encounter",
@@ -41,7 +43,6 @@
   "contactList": "Contact List",
   "contactTracingForm": "Contact tracing form",
   "contactType": "Contact Type",
-  "contractionLevel": "Contraction level",
   "date": "Date",
   "dateDate": "Date & time",
   "dateOfDelivery": "Date of Delivery",
@@ -52,7 +53,6 @@
   "deliveryDate": "Delivery Date",
   "deliveryOutcome": "Delivery Outcome",
   "department": "Department",
-  "descentOfHead": "Descent of Head",
   "disabilities": "Disabilities",
   "discard": "Discard",
   "dischargeDate": "Discharge Date",
@@ -81,18 +81,18 @@
   "failedDeleting": "couldn't be deleted",
   "familyContacts": "Family contacts",
   "feedingOrders": "Feeding Orders",
-  "fetalHeartRate": "Fetal Heart Rate",
+  "fieldRequired": "This field is required",
   "filterByForm": "Filter by form",
   "filterByStatue": "Filter by status",
   "filterByStatus": "Filter by status",
   "finalDiagnosis": "Final Diagnosis",
   "finalDischargeDiagnosis": "Final Discharge Diagnosis",
   "finalOutcome": "Final Outcome",
-  "foetalHeartRate": "Foetal Heart Rate",
   "foetalPresentation": "Foetal Presentation",
   "followUpDate": "Next follow-up date",
   "formName": "Form Name",
   "gestationalSize": "Gestational Size",
+  "graphView": "Graph View",
   "hivPatientSummary": "HIV Patient Summary",
   "hivStatus": "HIV Status",
   "hivTestResults": "HIV Status",
@@ -111,14 +111,15 @@
   "ipvOutCome": "IPV Outcome",
   "ipvQuestions": "IPV Questions",
   "labourAndDelivery": "Labour and Delivery",
+  "latest": "Latest",
   "listingDate": "Listing date",
   "livingWithClient": "Living with client",
   "loading": "Loading...",
+  "loadingData": "Loading data...",
   "managePeers": "Manage Peers",
   "maternalCondition": "Maternal Condition",
   "maternalSummary": "Maternal Summary",
   "medicalHistory": "Medical History",
-  "membraneAmnioticFluidAndMoulding": "Membrane Amniotic Fluid & Moulding",
   "missedAppointmentDate": "Date Missed Appointment",
   "modeOfDelivery": "Mode of Delivery",
   "motherGeneralCondition": "General condition",
@@ -129,14 +130,16 @@
   "nextVisitDate": "Next visit date",
   "no": "No",
   "noContactToDisplay": "There is no contact data to display for this patient.",
+  "noDataAvailable": "No data available for this graph",
   "noEncounters": "No encounters found",
   "noEncounterToDisplay": "There are no encounters to display for this patient.",
   "noObservationsFound": "No observations found",
+  "noPatientSelected": "No patient selected",
   "operatingDoctor": "Operating Doctor",
   "otherSubstanceAbuse": "Other Substance Abuse",
   "partnerStatus": "HIV status of partner)",
   "partograph": "Partograph",
-  "partographData": "Vital Components",
+  "partography": "Partography",
   "peerCalendar": "Peer Calendar",
   "pending": "Pending",
   "pendingDocumentations": "Pending Documentation",
@@ -166,8 +169,10 @@
   "save": "Save",
   "savedRelationship": "Relationship ended successfully",
   "selectClinic": "Select Clinic",
+  "selectPatientMessage": "Please select a patient to view partography data.",
   "sex": "Sex",
   "sexualAssault": "3.Has he/she ever forced you to do something sexually that made you feel uncomfortable?",
+  "showingResults": "Showing {{start}}-{{end}} of {{total}} {{itemType}}",
   "smoking": "Smoking",
   "smokingDuration": "Smoking Duration",
   "socialHistory": "Social History",
@@ -177,7 +182,7 @@
   "successfullyDeleted": "successfully deleted",
   "surgicalHistory": "Surgical History",
   "surgicalSummary": "Surgical Summary",
-  "tabList": "Tab List",
+  "tableView": "Table View",
   "tbScreening": "TB screening outcome",
   "testApproach": "Approach",
   "testDate": "Test Date",
@@ -187,7 +192,7 @@
   "testStrategy": "Strategy",
   "therapiesPrescribed": "Therapies Prescribed",
   "threatened": "2. Has he/she ever threatened to hurt you?",
-  "timeRecorded": "Time Recorded",
+  "time": "Time",
   "total": "Total",
   "totalPeers": "Total Peers",
   "traceContact": "Trace Contact",
@@ -200,5 +205,6 @@
   "viewBeds": "View beds",
   "viewEncounter": "View",
   "visitDate": "Screening Date",
+  "yearsOld": "years old",
   "yes": "Yes"
 }
